Handle disconnection of USB device automatically for Klipper firmware
Klipper firmware uses a virtual serial port at /tmp/printer which does
not disappear when USB device is lost.

This update enables the autoconnection and disconnection of Octoprint
from the virtual serial port.

Combined with the following GCODE script in octoprint, this facilitates
seamless connection/disconnection with Klipper:

"After connection to printer is established":
FIRMWARE_RESTART